提交用户管理模块

+++ src/components/User/User.vue
@@ -285,6 +285,15 @@ export default {
         })
       },
       async deleteUser(id){
+        const confirmResult=await this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).catch(err=>err)
+        console.log(confirmResult)
+        if(confirmResult=='cancel'){
+          return this.$message.info("已经取消删除")
+        }
         const {data:res}=await this.$http.delete('users/'+id)
         if(res.meta.status!=200)
           {
